Changed panel positions to correct locations
I'm having an issue with the bootstrap columns in the saves and stats panels.  If they add up to the full 12 column layout, the final column is pushed to the next line.  I'm not sure what is causing this behavior, but it wasn't doing that before switching to this layout.